FieldsController Split
