+++ lib/Alchemy/Phrasea/Application.php
@@ -12,7 +12,6 @@
 namespace Alchemy\Phrasea;
 
 use Alchemy\Geonames\GeonamesServiceProvider;
-use Alchemy\Phrasea\ControllerProvider\Admin\Fields;
 use Alchemy\Phrasea\ControllerProvider\Admin\Publications;
 use Alchemy\Phrasea\ControllerProvider\Admin\Root as AdminRoot;
 use Alchemy\Phrasea\ControllerProvider\Admin\SearchEngine;
@@ -312,6 +311,7 @@ class Application extends SilexApplication
             'Alchemy\Phrasea\ControllerProvider\Admin\Dashboard' => [],
             'Alchemy\Phrasea\ControllerProvider\Admin\Databox' => [],
             'Alchemy\Phrasea\ControllerProvider\Admin\Databoxes' => [],
+            'Alchemy\Phrasea\ControllerProvider\Admin\Fields' => [],
             'Alchemy\Phrasea\ControllerProvider\Admin\Users' => [],
             'Alchemy\Phrasea\ControllerProvider\Datafiles' => [],
             'Alchemy\Phrasea\ControllerProvider\Lightbox' => [],
@@ -621,7 +621,6 @@ class Application extends SilexApplication
         $this->mount('/admin/setup', new Setup());
         $this->mount('/admin/search-engine', new SearchEngine());
         $this->mount('/admin/publications', new Publications());
-        $this->mount('/admin/fields', new Fields());
         $this->mount('/admin/task-manager', new TaskManager());
         $this->mount('/admin/subdefs', new Subdefs());
 
@@ -671,6 +670,7 @@ class Application extends SilexApplication
             '/admin/dashboard'       => 'Alchemy\Phrasea\ControllerProvider\Admin\Dashboard',
             '/admin/databox'         => 'Alchemy\Phrasea\ControllerProvider\Admin\Databox',
             '/admin/databoxes'       => 'Alchemy\Phrasea\ControllerProvider\Admin\Databoxes',
+            '/admin/fields'          => 'Alchemy\Phrasea\ControllerProvider\Admin\Fields',
             '/admin/users'           => 'Alchemy\Phrasea\ControllerProvider\Admin\Users',
             '/datafiles'             => 'Alchemy\Phrasea\ControllerProvider\Datafiles',
             '/include/minify'        => 'Alchemy\Phrasea\ControllerProvider\Minifier',
